Rank and user security
Users with the permission Rank_Edit or Rank_Create need permission from
an Admin to create/edit a rank with permissions higher than what they
have.

Users with the permission User_Create or User_Edit need Admin permisson
to create/edit users with a higher rank that what they have.

When a admin confirms a invited user it will then send the email

Pending actions are displayed on the dashboard only for admins

Extra: Switched Ranks from being based of names to being bassed of ids