Custom settings for permissions and basic addon system
Addon manager and api created allow for adding to be created and for
them to interact with the pages(Check Wiki). Added custom settings for
permissions, .php file in /Permissons loaded under the permission.
Created custom elements that can be referenced(PageSelect.php). Created
custom settings for the Page rank, can chose what pages ranks are
allowed to edit