Mobile optimization
Index: Added a side navigation that slides in,
replaced the old slanted divs with a new one using transform so it was
responsive, changed the images so they were the create dimensions, got
rid of the CoderDojo icon in the top left, made the google maps iframe
responsive.

Blog.php: Added a side navigation that slides in.

Ninjas.php: Added a side navigation that slides in, improved the design
of the list version of the Ninja names so it was responsive and was
easier to read.

Scratch-projects.php: Made the project divs responsive, disabled the
ability to change the image into the iframe on devices that don’t have
flash(mobiles), added a side navigation that slides in from the side